Changed sharepoint handling from using Office365-REST-Python-Client to shareplum
