+++ .pylintrc
@@ -1,0 +1,7 @@
+[pylint.messages_control]
+disable = 
+  C0301, # Line too long
+  I1101, E1101, # C-modules members
+  R0913, # Too many arguments
+  R0914, # Too many local variables
+  W0718, # Catching too general exception Exception (broad-exception-caught)
