+++ pyproject.toml
@@ -17,7 +17,7 @@ classifiers = [
     "Operating System :: Microsoft :: Windows",
 ]
 dependencies = [
-  "Office365-REST-Python-Client >= 2.5.9",
+  "shareplum",
   "openpyxl >= 3.1.2",
   "pyodbc >= 5.1.0",
   "requests_ntlm >= 1.2.0",
